+++ CMakeLists.txt
@@ -760,6 +760,8 @@ add_library (seastar
   src/net/stack.cc
   src/net/tcp.cc
   src/net/tls.cc
+  src/net/tls-impl.cc
+  src/net/tls-impl.hh
   src/net/udp.cc
   src/net/unix_address.cc
   src/net/virtio.cc
